IOS-6526: CR fixes

+++ BlockchainSdk/Blockchains/Koinos/Models/KoinosAccountInfo.swift
@@ -9,6 +9,6 @@
 import Foundation
 
 struct KoinosAccountInfo {
-    let koinBalance: Decimal
-    let mana: Decimal
+    let koinBalance: UInt64
+    let mana: UInt64
 }
